# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DurianFarm AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List workers
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}:
    patch:
      operationId: updateWorker
      tags: [workers]
      summary: Update a worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
    delete:
      operationId: deleteWorker
      tags: [workers]
      summary: Delete a worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /attendance:
    get:
      operationId: listAttendance
      tags: [workers]
      summary: List attendance records
      parameters:
        - name: workerId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM"
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [workers]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    delete:
      operationId: deleteAttendance
      tags: [workers]
      summary: Delete an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Worker:
      type: object
      required: [id, name, role, status, hireDate, dailyWage, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
          format: date
        dailyWage:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WorkerInput:
      type: object
      required: [name, role, hireDate]
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
          format: date
        dailyWage:
          type: number
        notes:
          type: string
    WorkerUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
          format: date
        dailyWage:
          type: number
        notes:
          type: string
    Attendance:
      type: object
      required: [id, workerId, date, status, createdAt]
      properties:
        id:
          type: integer
        workerId:
          type: integer
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, leave, half_day]
        hoursWorked:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AttendanceInput:
      type: object
      required: [workerId, date, status]
      properties:
        workerId:
          type: integer
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, leave, half_day]
        hoursWorked:
          type: number
        notes:
          type: string
...