# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DurianFarm AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceStatus } from './attendanceStatus';

export interface Attendance {
  id: number;
  workerId: number;
  date: Date;
  status: AttendanceStatus;
  /** @nullable */
  hoursWorked?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}