# lib/api-zod/src/generated/types/workerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DurianFarm AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkerInputStatus } from './workerInputStatus';

export interface WorkerInput {
  name: string;
  role: string;
  phone?: string;
  status?: WorkerInputStatus;
  hireDate: Date;
  dailyWage?: number;
  notes?: string;
}